Use conventions provided by bundler gem command
