+++ google-id-token.gemspec
@@ -13,24 +13,32 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+lib = File.expand_path('../lib', __FILE__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+require 'google-id-token/version'
+
 Gem::Specification.new do |s|
   s.name = 'google-id-token'
-  s.version = '1.3.1'
+  s.version = GoogleIDToken::VERSION
 
   s.homepage = 'https://github.com/google/google-id-token/'
   s.license = 'APACHE-2.0'
   s.summary = 'Google ID Token utilities'
   s.description = 'Google ID Token utilities; currently just a parser/checker'
 
-  s.files = ['lib/google-id-token.rb', 'README.rdoc' ]
+  s.files         = `git ls-files -z`.split("\x0").reject do |f|
+    f.match(%r{^(test|spec|features)/})
+  end
 
   s.add_runtime_dependency 'multi_json'
   s.add_runtime_dependency 'jwt'
 
+  s.add_development_dependency 'bundler', '~> 1.13'
   s.add_development_dependency 'fakeweb'
   s.add_development_dependency 'rake'
   s.add_development_dependency 'rspec'
   s.add_development_dependency 'openssl'
+
 
   s.authors = ['Tim Bray', 'Bob Aman']
   s.email = 'tbray@textuality.com'
