+++ lib/google-id-token/version.rb
@@ -1,0 +1,3 @@
+module GoogleIDToken
+  VERSION = '1.3.1'
+end
